main: flicker free and new argument
The clearing of the screen after any step caused the screen to
flicker. This was not nice in higher resolutions. It is solved
now by just moving the cursor to the start position. New Areas
are drawn over the old one.

Also added a new flag to modify the delay between iterations
in milliseconds.